# src/main/resources/api/openapi.yaml
openapi: 3.1.0
info:
  title: api-document
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "3.1"
servers:
  - url: http://localhost:64339
    description: Generated server url
...
paths:
  /{municipalityId}/documents/{registrationNumber}/responsibilities:
    put:
      tags:
        - Documents
      summary: Replace document responsibilities.
      operationId: updateResponsibilities
      parameters:
        - name: municipalityId
          in: path
          description: Municipality ID
          required: true
          schema:
            type: string
          example: 2281
        - name: registrationNumber
          in: path
          description: Document registration number
          required: true
          schema:
            type: string
          example: 2023-2281-1337
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentResponsibilitiesUpdateRequest"
        required: true
      responses:
        "204":
          description: Successful operation
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/documents/{registrationNumber}/files:
    put:
      tags:
        - Documents
      summary: Add document file data (or replace existing if filename already exists
        on the document object). Accepts either a single file via 'documentFile' (deprecated)
        or multiple files via 'documentFiles' — in both cases the operation produces
        exactly one new revision.
      operationId: addOrReplaceFile
      parameters:
        - name: municipalityId
          in: path
          description: Municipality ID
          required: true
          schema:
            type: string
          example: 2281
        - name: registrationNumber
          in: path
          description: Document registration number
          required: true
          schema:
            type: string
          example: 2023-2281-1337
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                document:
                  $ref: "#/components/schemas/DocumentDataCreateRequest"
                  description: Document
                documentFile:
                  type: string
                  format: binary
                  description: Single file (deprecated — use 'documentFiles' for batch
                    upload).
                documentFiles:
                  type: array
                  items:
                    type: string
                    format: binary
              required:
                - document
      responses:
        "204":
          description: Successful operation
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/documents/{registrationNumber}/unrevoke:
    post:
      tags:
        - Documents
      summary: Unrevoke document. Targets latest revision by default; pass 'revision' to target a specific revision.
      operationId: unrevoke
      parameters:
        - name: municipalityId
          in: path
          required: true
          schema:
            type: string
        - name: registrationNumber
          in: path
          required: true
          schema:
            type: string
        - name: updatedBy
          in: query
          required: true
          schema:
            type: string
            minLength: 1
        - name: revision
          in: query
          description: Revision to transition. Defaults to the latest revision.
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
          example: 5
      responses:
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/documents/{registrationNumber}/revoke:
    post:
      tags:
        - Documents
      summary: Revoke document. Targets latest revision by default; pass 'revision' to target a specific revision.
      operationId: revoke
      parameters:
        - name: municipalityId
          in: path
          required: true
          schema:
            type: string
        - name: registrationNumber
          in: path
          required: true
          schema:
            type: string
        - name: updatedBy
          in: query
          required: true
          schema:
            type: string
            minLength: 1
        - name: revision
          in: query
          description: Revision to transition. Defaults to the latest revision.
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
          example: 5
      responses:
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/documents/{registrationNumber}/publish:
    post:
      tags:
        - Documents
      summary: Publish document. Targets latest revision by default; pass 'revision' to target a specific revision.
      operationId: publish
      parameters:
        - name: municipalityId
          in: path
          required: true
          schema:
            type: string
        - name: registrationNumber
          in: path
          required: true
          schema:
            type: string
        - name: updatedBy
          in: query
          required: true
          schema:
            type: string
            minLength: 1
        - name: revision
          in: query
          description: Revision to transition. Defaults to the latest revision.
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
          example: 5
      responses:
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/documents/filter:
    post:
      tags:
        - Documents
      summary: Search documents by parameters
      description: "Parameters:\n- createdBy: Filter by the user that created the\
        \ document. Datatype - String\n- includeConfidential: Should the search include\
        \ confidential documents? Datatype - boolean (default: false)\n- onlyLatestRevision:\
        \ Should the search include only the latest revision of the documents? Datatype\
        \ - boolean (default: false)\n- documentTypes: Which document types to include\
        \ in the search. Datatype - List of Strings\n- statuses: Which lifecycle statuses\
        \ to include. Datatype - List of DocumentStatus. If omitted, defaults to published\
        \ statuses (SCHEDULED, ACTIVE, EXPIRED) - DRAFT and REVOKED are excluded.\
        \ When set explicitly, the list is used as-is (admin may pass [DRAFT] to see\
        \ only drafts).\n- metaData: Uses the metadata object to search for documents\
        \ with specific metadata. Datatype - List of metadata objects.\n- responsibilities:\
        \ Uses document responsibilities to search for documents where at least one\
        \ personId matches. Datatype - List of DocumentResponsibility objects.\n- page:\
        \ The page number\
        \ to retrieve. Datatype - integer (default: 1)\n- limit: The number of documents\
        \ to retrieve per page. Datatype - integer (default: 100)\n\nObjects:\n- MetaData:\
        \ {\n\t- key: A given metadata key, this is optional. All metadata will be\
        \ searched if key is not provided. Datatype - String\n\t- matchesAny: Returns\
        \ documents where metadata entry with the given key have at least one of the\
        \ matchesAny values (if key is present), or if the complete set of metadata\
        \ have at least one of the matchesAny (when no key is present). Datatype -\
        \ List of Strings\n\t- matchesAll: Returns documents where metadata entry\
        \ with the given key have at least one of the matchesAny values (if key is\
        \ present), or if the complete set of metadata have at least one of the matchesAny\
        \ (when no key is present). Datatype - List of Strings\n}\n- DocumentResponsibility:\
        \ {\n\t- personId: Person ID of the responsible party. Datatype - String (UUID)\n}\n"
      operationId: searchByParameters
      parameters:
        - name: municipalityId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentParameters"
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedDocumentResponse"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/documents/{registrationNumber}/confidentiality:
    patch:
      tags:
        - Documents
      summary: Update document confidentiality (on all revisions).
      operationId: updateConfidentiality
      parameters:
        - name: municipalityId
          in: path
          description: Municipality ID
          required: true
          schema:
            type: string
          example: 2281
        - name: registrationNumber
          in: path
          description: Document registration number
          required: true
          schema:
            type: string
          example: 2023-2281-1337
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfidentialityUpdateRequest"
        required: true
      responses:
        "200":
          description: Successful operation
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
    Problem:
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
    DocumentResponsibilitiesUpdateRequest:
      type: object
      description: Document responsibilities update request model.
      properties:
        updatedBy:
          type: string
          description: PersonId of the actor that performed this change.
          examples:
            - 6c3e4f5a-7b8d-4e9c-a1f2-d3e4b5c6a7f8
          minLength: 1
        responsibilities:
          type: array
          items:
            $ref: "#/components/schemas/DocumentResponsibility"
            description: List of document responsibilities.
      required:
        - updatedBy
        - responsibilities
...
    ConstraintViolationProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        detail:
          type: string
        causeAsProblem:
          $ref: "#/components/schemas/ThrowableProblem"
        instance:
          type: string
          format: uri
...
    DocumentDataCreateRequest:
      type: object
      description: DocumentDataCreateRequest model.
      properties:
        createdBy:
          type: string
          description: PersonId of the actor that created this revision.
          examples:
            - 6c3e4f5a-7b8d-4e9c-a1f2-d3e4b5c6a7f8
          minLength: 1
      required:
        - createdBy
...
  securitySchemes: {}